refactor(dotcom): extract shared thumbnail render core from MCP route
The OG image route and queue consumer imported their rate limiting,
Browser Rendering invocation, snapshot loading, and page enumeration
from the MCP route module, and each surface kept its own near-identical
board resolver and telemetry writer that had to stay in sync by hand.

Move the render-and-cache core into routes/tla/thumbnailRender.ts:

- one resolveThumbnailBoard with the shared public-view gates;
  resolveSharedBoardById and the OG surfaces are now thin adapters
- one writeScreenshotTelemetry writer for the shared dataset layout
- one captureThumbnailScreenshot helper for the mint-token-and-render
  sequence, plus a putThumbnailPng cache write both surfaces use
- a ThumbnailBoardKind type replacing eight inline spellings of
  'published' | 'shared_file'

Also move the chunked base64 helpers to utils/base64.ts (renderTokens
kept its own copies), move getPublicOrigin out of the getOgImage route
into utils/, drop the OG_IMAGE_WIDTH/HEIGHT aliases, derive the OG
pending-marker key from the cache key, and share the room-snapshot
builder across the screenshot test files via screenshotTestHelpers.

No behavior change; the sync-worker tests cover both surfaces.
